Fix handing of pgsql arrays for the literal and argument-binding cases.
For PostgreSQL arrays, strings must be escaped differently depending on
they are used directly into SQL statements or if they are used in, say,
a parameterized SQL statement with argument binding.  In both cases,
we need to escape " and \ as \" and \\, but we also need to escape
' as '' in the non binding case.