Add EAN13 data type support via isn extension
This commit is my first attempt at integrating the isn extension with
postgres_ext. I added EAN13 first because it is the superset of the
other included types. The style, structure, and syntax are mostly
parroted from postgres_ext's existing data types. As such, there are a
few extraneous (green) specs as well as room for optimization and
improvement.