Fix error when creating a join on STI relationship
The visit_Array method in the Arel < version 4.1 PostgreSQL visitor
tries to query the schema cache to get the type the column. When querying
with an STI model using a JOIN statement the Arel node that gets passed
to this method is an Arel::Nodes::TableAlias with a name that is not in
the connection's schema cache. Attempting to query the schema cache results
in a database error.

This workaround simply tests if the relation type passed to the
visit_Array method is nil or a Arel::Nodes::TableAlias and, if so, does
not query the schema cache for the column type.

Closes #154