Reset index opclasses after add_extension
This fixes an issue in rails 3.2 where add_extension can cause the
schema dumper to create a malformed schema.rb.

Some extensions like pg_trgm and hstore modify the available index
opclasses in the pg_opclass table, but the query to this table is
memoized in ActiveRecord::Connection.base#opclasses for the entire
lifetime of a connection, even though correctly determining the opclass
when dumping to schema.rb relies on the updated table.  The result is
that a migration like

    add_extension('pg_trgm')
    add_index "items", "col",
      :name => "idx", :using => :gin, :index_opclass => :gin_trgm_ops

will generate a schema.rb that excludes the :index_opclass option:

    add_index "items", ["col"], :name => "idx", :using => :gin

This is malformed since a gin index on a string column is illegal
without an opclass, so running something that uses schema.rb such as
rake db:test:prepare results in the following postgres error:

    PG::UndefinedObject: ERROR:  data type character varying has no
    default operator class for access method "gin"
    HINT:  You must specify an operator class for the index or define
    a default operator class for the data type.

Clearing the memoized variable after adding any extensions fixes this issue.