add ability to build with statement from raw arel
This is particularly important when trying to to do bulk updates since
we need to create a table with a defined schema e.g
"WITH update(id,new_value)", and that's not currently supported.
Implemented by changing the case statement in with build to return
with_statement of type Arel::Nodes::As.